ml_core/requirements.txt
torch>=2.2.0
torch-geometric>=2.5.0
pytorch-lightning>=2.1.0  # Advanced Training Loop
hydra-core>=1.3.0         # Configuration Management
wandb>=0.16.0             # Experiment Tracking
protobuf>=4.25.0          # Binary Serialization
lmdb>=1.4.1               # Fast Database
transformers>=4.38.0
pandas
scikit-learn
tqdm
